Major new version
This version brings statistics upload and many bugfixes.

+++ src/org/pixmob/freemobile/netstat/ui/BatteryChart.java
@@ -60,32 +60,28 @@ public class BatteryChart extends View {
     private float textCursorBottom;
     private String textCursor;
     private int textCursorFormat = DateUtils.FORMAT_SHOW_TIME;
-    
+
     public BatteryChart(final Context context, final AttributeSet attrs) {
         super(context, attrs);
     }
-    
+
     @Override
     protected void onDetachedFromWindow() {
         super.onDetachedFromWindow();
         invalidateCache();
     }
-    
+
     @Override
     public boolean onTouchEvent(MotionEvent event) {
         final int action = event.getAction();
-        if (MotionEvent.ACTION_DOWN == action
-                || MotionEvent.ACTION_MOVE == action) {
+        if (MotionEvent.ACTION_DOWN == action || MotionEvent.ACTION_MOVE == action) {
             final float x = event.getX();
             if (x >= graphLeft && x <= graphRight) {
                 if (events != null && events.length != 0) {
                     final long t0 = events[0].timestamp;
-                    final long t = (long) Math.round((x - graphLeft)
-                            / getXFactor())
-                            + t0;
-                    textCursor = DateUtils.formatDateTime(getContext(), t,
-                        textCursorFormat);
-                    
+                    final long t = Math.round((x - graphLeft) / getXFactor()) + t0;
+                    textCursor = DateUtils.formatDateTime(getContext(), t, textCursorFormat);
+
                     touchX = x;
                     invalidate();
                     return true;
@@ -99,19 +95,18 @@ public class BatteryChart extends View {
         }
         return super.onTouchEvent(event);
     }
-    
+
     private float getXFactor() {
         if (events == null || events.length == 0) {
             return 0;
         }
-        
-        return (graphRight - graphLeft)
-                / (events[events.length - 1].timestamp - events[0].timestamp);
+
+        return (graphRight - graphLeft) / (events[events.length - 1].timestamp - events[0].timestamp);
     }
-    
+
     private void doDraw(Canvas canvas) {
         final float y0Mob = 0;
-        
+
         // Draw every lines with a single call to Canvas#drawLines, for better
         // performance.
         final float[] lines = new float[12 * 4];
@@ -124,7 +119,7 @@ public class BatteryChart extends View {
         lines[lineIdx++] = graphBottom;
         lines[lineIdx++] = graphRight;
         lines[lineIdx++] = graphBottom;
-        
+
         // Draw Y units.
         final float x0Text = graphLeft - toDip(5);
         final float yFactor = (graphBottom - graphTop) / 100f;
@@ -140,29 +135,28 @@ public class BatteryChart extends View {
             final float y = graphBottom - yFactor * i;
             canvas.drawRect(graphLeft, y, graphRight, y + yBand, bgPaint);
             if (i != 100) {
-                canvas.drawText(String.valueOf(i), x0Text, y - yAscent2,
-                    yTextPaint);
+                canvas.drawText(String.valueOf(i), x0Text, y - yAscent2, yTextPaint);
             }
             lines[lineIdx++] = graphLeft;
             lines[lineIdx++] = y;
             lines[lineIdx++] = graphRight;
             lines[lineIdx++] = y;
         }
-        
+
         final boolean drawChart = events != null && events.length > 1;
         Path batteryLevelBorderPath = null;
-        
+
         if (drawChart) {
             final long t0 = events[0].timestamp;
             final float xFactor = getXFactor();
-            
+
             final int eventCount = events.length;
-            
+
             final Path batteryPath = new Path();
             batteryPath.moveTo(graphLeft, graphBottom);
             batteryPath.incReserve(eventCount + 2);
             float lastY = 0;
-            
+
             for (int i = 0; i < eventCount; ++i) {
                 final Event e = events[i];
                 float x = (e.timestamp - t0) * xFactor + graphLeft;
@@ -170,16 +164,15 @@ public class BatteryChart extends View {
                     continue;
                 }
                 final float y = graphBottom - e.batteryLevel * yFactor;
-                
+
                 if (i != 0) {
                     batteryPath.lineTo(x, y);
                 } else {
                     batteryPath.moveTo(x, y);
                 }
                 lastY = y;
-                
-                final MobileOperator mobOp = MobileOperator
-                        .fromString(e.mobileOperator);
+
+                final MobileOperator mobOp = MobileOperator.fromString(e.mobileOperator);
                 if (mobOp != null) {
                     float x2 = graphLeft;
                     if (i != 0) {
@@ -187,16 +180,15 @@ public class BatteryChart extends View {
                         if (e.powerOn && !e0.powerOn) {
                             continue;
                         }
-                        
-                        final MobileOperator mobOp0 = MobileOperator
-                                .fromString(e0.mobileOperator);
+
+                        final MobileOperator mobOp0 = MobileOperator.fromString(e0.mobileOperator);
                         if (mobOp0 != null) {
                             if (MobileOperator.FREE_MOBILE.equals(mobOp0)) {
                                 mobileOperatorPaint.setColor(freeMobileColor);
                             } else if (MobileOperator.ORANGE.equals(mobOp0)) {
                                 mobileOperatorPaint.setColor(orangeColor);
                             }
-                            
+
                             x2 = (e0.timestamp - t0) * xFactor + graphLeft;
                             if (x2 < graphLeft) {
                                 x2 = graphLeft;
@@ -204,45 +196,42 @@ public class BatteryChart extends View {
                             if (x2 > graphRight) {
                                 x2 = graphRight;
                             }
-                            canvas.drawLine(x2, y0Mob, x, y0Mob,
-                                mobileOperatorPaint);
+                            canvas.drawLine(x2, y0Mob, x, y0Mob, mobileOperatorPaint);
                         }
                     }
                 }
             }
-            
+
             batteryLevelBorderPath = new Path(batteryPath);
-            
+
             batteryPath.lineTo(graphRight, lastY);
             batteryPath.lineTo(graphRight, graphBottom);
             batteryPath.lineTo(graphLeft, graphBottom);
             canvas.drawPath(batteryPath, batteryLevelPaint);
         }
-        
+
         // Draw axes.
         canvas.drawLines(lines, yBarPaint);
-        
+
         if (drawChart) {
             canvas.clipRect(graphLeft, 0, graphRight, graphBottom);
             canvas.drawPath(batteryLevelBorderPath, batteryLevelBorderPaint);
         }
     }
-    
+
     @Override
     protected void onDraw(Canvas canvas) {
         super.onDraw(canvas);
-        
+
         // Lazy initialize paint properties, once.
         if (batteryLevelPaint == null) {
             batteryLevelPaint = new Paint();
             batteryLevelPaint.setStyle(Paint.Style.FILL);
-            
-            final int c1 = getResources()
-                    .getColor(R.color.battery_level_color1);
-            final int c2 = getResources()
-                    .getColor(R.color.battery_level_color2);
-            batteryLevelPaint.setShader(new LinearGradient(0, 0, 0,
-                    getHeight(), c1, c2, Shader.TileMode.CLAMP));
+
+            final int c1 = getResources().getColor(R.color.battery_level_color1);
+            final int c2 = getResources().getColor(R.color.battery_level_color2);
+            batteryLevelPaint.setShader(new LinearGradient(0, 0, 0, getHeight(), c1, c2,
+                    Shader.TileMode.CLAMP));
         }
         if (batteryLevelBorderPaint == null) {
             batteryLevelBorderPaint = new Paint();
@@ -251,8 +240,7 @@ public class BatteryChart extends View {
             batteryLevelBorderPaint.setStrokeWidth(4);
             batteryLevelBorderPaint.setStrokeCap(Paint.Cap.ROUND);
             batteryLevelBorderPaint.setStrokeJoin(Paint.Join.ROUND);
-            batteryLevelBorderPaint.setColor(getResources().getColor(
-                R.color.battery_level_border_color));
+            batteryLevelBorderPaint.setColor(getResources().getColor(R.color.battery_level_border_color));
         }
         if (bgPaint == null) {
             bgColor1 = getResources().getColor(R.color.battery_bg_color1);
@@ -264,36 +252,30 @@ public class BatteryChart extends View {
             yTextPaint = new Paint();
             yTextPaint.setAntiAlias(true);
             yTextPaint.setSubpixelText(true);
-            yTextPaint.setColor(getResources().getColor(
-                R.color.battery_y_text_color));
-            yTextPaint.setTextSize(getResources().getDimension(
-                R.dimen.battery_y_text_size));
+            yTextPaint.setColor(getResources().getColor(R.color.battery_y_text_color));
+            yTextPaint.setTextSize(getResources().getDimension(R.dimen.battery_y_text_size));
             yTextPaint.setTextAlign(Paint.Align.RIGHT);
             yTextPaint.setStrokeWidth(2);
-            
+
             graphLeft = yTextPaint.measureText("100") + 2;
         }
         if (yBarPaint == null) {
             yBarPaint = new Paint();
-            yBarPaint.setColor(getResources().getColor(
-                R.color.battery_y_bar_color));
+            yBarPaint.setColor(getResources().getColor(R.color.battery_y_bar_color));
             yBarPaint.setStrokeWidth(2);
         }
         if (mobileOperatorPaint == null) {
             mobileOperatorPaint = new Paint();
             mobileOperatorPaint.setStrokeWidth(16);
             mobileOperatorPaint.setStyle(Paint.Style.STROKE);
-            
-            orangeColor = getResources()
-                    .getColor(R.color.orange_network_color1);
-            freeMobileColor = getResources().getColor(
-                R.color.free_mobile_network_color1);
+
+            orangeColor = getResources().getColor(R.color.orange_network_color1);
+            freeMobileColor = getResources().getColor(R.color.free_mobile_network_color1);
         }
         if (cursorPaint == null) {
             cursorPaint = new Paint();
             cursorPaint.setAntiAlias(true);
-            cursorPaint.setColor(getResources().getColor(
-                R.color.battery_cursor_color));
+            cursorPaint.setColor(getResources().getColor(R.color.battery_cursor_color));
             cursorPaint.setStrokeWidth(2);
             cursorPaint.setStyle(Paint.Style.FILL);
         }
@@ -301,61 +283,55 @@ public class BatteryChart extends View {
             textCursorPaint = new Paint();
             textCursorPaint.setAntiAlias(true);
             textCursorPaint.setSubpixelText(true);
-            textCursorPaint.setColor(getResources().getColor(
-                R.color.battery_text_cursor_color));
-            textCursorPaint.setTextSize(getResources().getDimension(
-                R.dimen.battery_y_text_size));
+            textCursorPaint.setColor(getResources().getColor(R.color.battery_text_cursor_color));
+            textCursorPaint.setTextSize(getResources().getDimension(R.dimen.battery_y_text_size));
             textCursorPaint.setTextAlign(Paint.Align.RIGHT);
-            
+
             textCursorBottom = -textCursorPaint.ascent() + toDip(5);
             graphTop = textCursorBottom + toDip(4);
         }
-        
+
         // No image is available in the cache: render a new one.
         if (cache == null) {
-            cache = Bitmap.createBitmap(getWidth(), getHeight(),
-                Bitmap.Config.ARGB_8888);
+            cache = Bitmap.createBitmap(getWidth(), getHeight(), Bitmap.Config.ARGB_8888);
             final Canvas c = new Canvas(cache);
             doDraw(c);
         }
-        
+
         canvas.drawBitmap(cache, 0, 0, null);
-        
+
         if (touchX != -1) {
             canvas.drawLine(touchX, graphTop, touchX, graphBottom, cursorPaint);
-            canvas.drawText(textCursor, textCursorLeft, textCursorBottom,
-                textCursorPaint);
+            canvas.drawText(textCursor, textCursorLeft, textCursorBottom, textCursorPaint);
         }
     }
-    
+
     @Override
     protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
-        setMeasuredDimension(MeasureSpec.getSize(widthMeasureSpec),
-            MeasureSpec.getSize(heightMeasureSpec));
+        setMeasuredDimension(MeasureSpec.getSize(widthMeasureSpec), MeasureSpec.getSize(heightMeasureSpec));
     }
-    
+
     @Override
     protected void onSizeChanged(int w, int h, int oldw, int oldh) {
         super.onSizeChanged(w, h, oldw, oldh);
         invalidateCache();
-        
+
         graphRight = getWidth();
         graphBottom = getHeight();
         textCursorLeft = graphRight - toDip(2);
     }
-    
+
     private float toDip(float pixels) {
-        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, pixels,
-            getResources().getDisplayMetrics());
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, pixels, getResources()
+                .getDisplayMetrics());
     }
-    
+
     public void setData(Event[] events) {
         this.events = events;
         invalidateCache();
-        
+
         if (events != null && events.length != 0) {
-            final long duration = events[events.length - 1].timestamp
-                    - events[0].timestamp;
+            final long duration = events[events.length - 1].timestamp - events[0].timestamp;
             if (duration >= 86400 * 1000) {
                 // Over a day: show date+time.
                 textCursorFormat = DateUtils.FORMAT_SHOW_DATE;
@@ -365,7 +341,7 @@ public class BatteryChart extends View {
             }
         }
     }
-    
+
     private void invalidateCache() {
         if (cache != null) {
             cache.recycle();
